Move navigation to currency selection to currencies hook
