+++ src/helpers/currencySelectionTypes.js
@@ -1,0 +1,4 @@
+export default {
+  input: 'input',
+  output: 'output',
+};
